refactor/indexserver: isolate clone lifetime before adding reuse
Keep Git acquisition and cleanup out of gitIndex so it only orchestrates
configuration and indexing. Group the existing Git operations under
gitRepoCache and lend the fetched directory through a callback.

This deliberately preserves temporary-clone behavior, including Git commands,
delta fallback, and unconditional best-effort deletion. Keeping this separate
from retention policy makes the upcoming monorepo cache change reviewable.

Amp-Thread-ID: https://ampcode.com/threads/T-01a0717a-3acd-7026-977e-6ffe6d30e027
